+++ uCOS-III_Wrapper/uCOS-III/os_flag.c
@@ -491,6 +491,7 @@ OS_FLAGS  OSFlagPend (OS_FLAG_GRP  *p_grp,
 
     CPU_CRITICAL_ENTER();
     OSTCBCurPtr->PendStatus = OS_STATUS_PEND_OK;            /* Clear pend status                                      */
+    OSTCBCurPtr->TaskState = OS_TASK_STATE_PEND;
     CPU_CRITICAL_EXIT(); 
     
     rt_err = rt_event_recv(&p_grp->FlagGrp,
@@ -719,7 +720,8 @@ OS_FLAGS  OSFlagPost (OS_FLAG_GRP  *p_grp,
                       OS_ERR       *p_err)
 {
     rt_err_t rt_err;
-        
+    struct rt_thread *thread;
+    
 #ifdef OS_SAFETY_CRITICAL
     if (p_err == (OS_ERR *)0) {
         OS_SAFETY_CRITICAL_EXCEPTION();
@@ -763,6 +765,13 @@ OS_FLAGS  OSFlagPost (OS_FLAG_GRP  *p_grp,
     
     rt_err = rt_event_send(&p_grp->FlagGrp,flags);
     *p_err = rt_err_to_ucosiii(rt_err);
+    if(rt_err == RT_EOK)
+    {
+        /*获取当前等待Flag的线程*/
+        thread = rt_list_entry(p_grp->FlagGrp.parent.suspend_thread.next, struct rt_thread, tlist);
+        ((OS_TCB*)thread)->TaskState = OS_TASK_STATE_RDY;/*更新任务状态*/
+    }
+    
     return p_grp->FlagGrp.set;/*返回执行后事件标志组的值*/
 }