+++ uCOS-III_Wrapper/uCOS-III/os_sem.c
@@ -419,6 +419,7 @@ OS_SEM_CTR  OSSemPend (OS_SEM   *p_sem,
     
     CPU_CRITICAL_ENTER();
     OSTCBCurPtr->PendStatus = OS_STATUS_PEND_OK;            /* Clear pend status                                      */
+    OSTCBCurPtr->TaskState = OS_TASK_STATE_PEND;
     CPU_CRITICAL_EXIT(); 
     
     rt_err = rt_sem_take(&p_sem->Sem,time);
@@ -588,6 +589,7 @@ OS_SEM_CTR  OSSemPost (OS_SEM  *p_sem,
                        OS_ERR  *p_err)
 {
     rt_err_t rt_err;
+    struct rt_thread *thread;
     
 #ifdef OS_SAFETY_CRITICAL
     if (p_err == (OS_ERR *)0) {
@@ -627,9 +629,16 @@ OS_SEM_CTR  OSSemPost (OS_SEM  *p_sem,
     {
         case OS_OPT_POST_1:
             rt_err = rt_sem_release(&p_sem->Sem);
+            if(rt_err == RT_EOK)
+            {
+                /*获取当前等待sem的线程*/
+                thread = rt_list_entry(p_sem->Sem.parent.suspend_thread.next, struct rt_thread, tlist);
+                ((OS_TCB*)thread)->TaskState = OS_TASK_STATE_RDY;/*更新任务状态*/
+            }
             break;
         
         case OS_OPT_POST_ALL:
+            /*在rt_sem_release_all内部已经更新TaskStatus*/
             rt_err = rt_sem_release_all(&p_sem->Sem);
             break;